requirements.txt
gTTS==2.5.1
pandas==2.2.1
pillow==10.2.0
moviepy==1.0.3
assemblyai==0.23.0
requests==2.31.0